website
connected all pages to one another. The template page is on as the projects page for now. All that is left is the projects page, which is three projects: two previous ones from ITI/DCIM and this one being sandwhiched together by the other two. THe style of the projects is like an infographic template.